mongodb-performance-test/requirements.txt
locust
pymongo[srv]
python-dotenv